Track cells from frame larger than one
In response to:  The problem is that I have some movies where I only
want to start analyzing at some point after frame 1. I only added
breakpoints from when I wanted it analyzed, but when I tell MMJ to
track, it automatically moves the first cell to frame 1. Since not all
cells start on the same frame, simply deleting the frames in front does
not work.